Move ResourceGetter to common
